Relocate select-k implementation
